backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.20
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic[email]==2.10.4
pydantic-settings==2.7.0
pandas==2.2.3
openpyxl==3.1.5